Added Repositories Added Functionality in REST-Method for adding Card and Deck tweaked model structure slightly

+++ src/main/java/Model/Battle/Round.java
@@ -8,16 +8,16 @@ public class Round {
 
     RoundOutcome roundOutcome;
 
-    int winnerDmg;
-    int looserDmg;
+    double winnerDmg;
+    double looserDmg;
 
     ACard winner;
     ACard looser;
 
     public void fight(ACard attackingCard,ACard defendingCard)
     {
-        int attackingCardDmg=0;
-        int defendingCardDmg=0;
+        double attackingCardDmg=0;
+        double defendingCardDmg=0;
 
         attackingCardDmg = attackingCard.calcDamage(State.ATTACK,defendingCard);
         defendingCardDmg = defendingCard.calcDamage(State.DEFEND,attackingCard);
